Roll forward to latest Okta starter and SDK.
Now that we are on Spring 2.4.x, we can have the latest Okta SDK again.

This reverts commit cb2dca821774f6b818f1044a9e8ff13362804012.

+++ backend/src/test/java/gov/cdc/usds/simplereport/service/BaseServiceTest.java
@@ -20,6 +20,7 @@ import com.okta.sdk.client.Client;
 import com.okta.sdk.client.Clients;
 import com.okta.sdk.resource.user.User;
 import com.okta.sdk.resource.group.Group;
+import com.okta.sdk.resource.group.GroupType;
 
 import gov.cdc.usds.simplereport.test_util.DbTruncator;
 import gov.cdc.usds.simplereport.test_util.SliceTestConfiguration;
@@ -107,7 +108,8 @@ public abstract class BaseServiceTest<T> {
     protected void clearOktaGroups() {
         for (Group g : _oktaClient.listGroups()) {
             String groupName = g.getProfile().getName();
-            if (groupName.startsWith(_authorizationProperties.getRolePrefix())) {
+            if (g.getType() == GroupType.OKTA_GROUP &&
+                        groupName.startsWith(_authorizationProperties.getRolePrefix())) {
                 g.delete();
             }
         }
